broke page tempaltes out for better readibility

+++ signals.py
@@ -23,16 +23,25 @@ def dataresource_post_save(sender, instance, *args, **kwargs):
 
 
 def purge_cache_on_delete(sender, instance, *args, **kwargs):
-    instance.resource.clear_cache()
-    if instance.resource_file and os.path.exists(os.path.join(s.MEDIA_ROOT, instance.resource_file.name)):
-        os.unlink(os.path.join(s.MEDIA_ROOT, instance.resource_file.name))
+    if sender is DataResource:
+        instance.resource.clear_cache()
+        if instance.resource_file and os.path.exists(os.path.join(s.MEDIA_ROOT, instance.resource_file.name)):
+            os.unlink(os.path.join(s.MEDIA_ROOT, instance.resource_file.name))
+    elif sender is Style:
+        for layer in instance.renderedlayer_set.all():
+            layer.resource.clear_cache()
+    elif sender is RenderedLayer:
+        layer.resource.clear_cache()
+
 
 def purge_resource_data(sender, instance, *args, **kwargs):
     if sender is DataResource:
         instance.driver_instance.clear_cache()
 
+
 def shave_tile_caches(sender, instance, bbox, *args, **kwargs):
     drivers.CacheManager.get().shave_caches(instance, bbox)
+
 
 def trim_tile_caches(sender, instance, *args, **kwargs):
     if sender is Style:
